refactor(#906): decompose DetailsScreen.kt (1739 → 656 lines)
Extract three logical sections into focused files:

MangaHeader.kt (449 lines)
  - MangaHeader composable: hero cover, parallax scroll, content-type
    detection, bloom animation, metadata chips, and action buttons
  - renderMarkdown: AnnotatedString renderer for bold/italic/links
  - Private regex and parallax constants scoped to this file

MangaSettingsSection.kt (512 lines)
  - DeleteAfterReadOption, NotificationOption
  - ReaderSettingsSection with all sub-composables: DirectionOption,
    PreloadOption, ReaderModeOption, ColorFilterOption,
    CustomTintColorPicker, BackgroundColorPicker, ColorChip

ChapterSection.kt (217 lines)
  - ChapterListHeader (simple variant used by detailsChapterItems)
  - ChapterFilterDialog with read/bookmark/download/scanlator filters
  - TriStateRow, EmptyScreen

DetailsScreen.kt retains the route composable, two-pane layout
coordinator, lazy-column section extensions, notes/description/
suggestions composables, and the @Preview.

All moved functions changed from private (file-scoped) to internal
(module-scoped). No logic changes.
